Extract `FollowStatus` into a separate file
- Remove webpack `resolveOptions` config for crypto
- Imported `FollowStatus` from new file in client bundle
- Extracted `FollowStatus` from `follow.tsx`
- Simplified `follow.tsx`
- Added `followStatus.tsx`
