+++ webpack.config.ts
@@ -64,7 +64,6 @@ function resolve(
 	// We rely on Buffer for our bridget thrift client
 	if (isClient) {
 		resolveOptions.alias = { ...resolveOptions.alias, Buffer: 'buffer' };
-		resolveOptions.fallback = { crypto: false };
 	}
 
 	return resolveOptions;
